+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/azure_cli.py
@@ -32,6 +32,10 @@ class AzureCliCredential(AsyncContextManager):
 
     This requires previously logging in to Azure via "az login", and will use the CLI's currently logged in identity.
     """
+    def __init__(self, tenant_id: str = ""):
+        AsyncContextManager.__init__(self)
+
+        self.tenant_id = tenant_id
 
     @log_get_token_async
     async def get_token(self, *scopes: str, **kwargs: "Any") -> "AccessToken":
@@ -55,7 +59,8 @@ class AzureCliCredential(AsyncContextManager):
 
         resource = _scopes_to_resource(*scopes)
         command = COMMAND_LINE.format(resource)
-        tenant = resolve_tenant("", **kwargs)
+        tenant = resolve_tenant(default_tenant= self.tenant_id, **kwargs)
+
         if tenant:
             command += " --tenant " + tenant
         output = await _run_command(command)