+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## 1.11.0b3 (Unreleased)
 
+Azure-identity is supported on Python 3.7 or later. For more details, please read our page on [Azure SDK for Python version support policy](https://github.com/Azure/azure-sdk-for-python/wiki/Azure-SDKs-Python-version-support-policy).
+
 ### Features Added
+
+- Added ability to specify `tenant_id` for `AzureCliCredential` (thanks @tikicoder)    ([#25207](https://github.com/Azure/azure-sdk-for-python/pull/25207))
 
 ### Breaking Changes